refactor(structure): components 삭제 및 폴더 구조 개편
