+++ src/app/routes/auth/register.tsx
@@ -1,7 +1,6 @@
-import type { Route } from "../+types/landing";
 import RegisterPage from "~pages/auth/register";
 
-export function meta(_: Route.MetaArgs) {
+export function meta() {
   return [
     { title: "Discord Clone - 회원가입" },
     { name: "description", content: "회원가입 중입니다." },
