+++ src/entities/notification/ui/icons/at-icon.tsx
