+++ src/entities/notification/ui/icons/filter-icon.tsx
