+++ src/entities/notification/ui/icons/friends-icon.tsx
