+++ src/entities/notification/ui/icons/mail-box-icon.tsx
