+++ src/entities/notification/ui/icons/satellite-icon.tsx
