+++ src/entities/notification/ui/icons/support-icon.tsx
