+++ src/entities/notification/ui/send-mail-box.tsx
