+++ src/entities/notification/ui/tabs-slot.tsx
