+++ src/features/auth/kakao/lib/redirect-to-kakao-auth.ts
