--- src/features/auth/login/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./kakao-loign-button";
