+++ src/widgets/landing/ui/navigation-list.tsx
