+++ tsconfig.json
@@ -17,7 +17,6 @@
     "paths": {
       "~/*": ["src/*"],
       "~app/*": ["src/app/*"],
-      "~components/*": ["src/components/*"],
       "~entities/*": ["src/entities/*"],
       "~features/*": ["src/features/*"],
       "~pages/*": ["src/pages/*"],
